+++ Isolator.Fody/IsolatorExecute.ContextName.cs
@@ -1,0 +1,72 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using Mono.Cecil;
+using Mono.Cecil.Cil;
+
+public partial class ModuleWeaver
+{
+    List<string> _isolatorContextNames;
+    MethodDefinition _setContextName;
+    private void InjectSetContextName(MethodDefinition method, CustomAttribute isolatorCustomAttribute)
+    {
+        _isolatorContextNames ??= GetListOfUniqueIsolatorConstextNames();
+        _setContextName ??= _targetType?.Methods.SingleOrDefault(_ => _.Name == "SetContextName");
+
+        // Only inject SetContextName if there are any custom context names defined
+        if (_isolatorContextNames.Count > 0)
+        {
+            var contextNameValue = isolatorCustomAttribute?.ConstructorArguments[0].Value as string;
+            InjectMethodWithStringParameter(method, _setContextName, contextNameValue);
+        }
+    }
+
+
+    private List<string> GetListOfUniqueIsolatorConstextNames()
+    {
+        var isolatorCustomAttributes = ModuleDefinition.GetTypes()
+            .Where(IsValidIsolatorTypeDefinition)
+            .Select(e => e.CustomAttributes
+            .SingleOrDefault(x => x.AttributeType.FullName == IsolatorAttribute))
+            .OfType<CustomAttribute>();
+
+        var isolatorCustomValues = isolatorCustomAttributes
+            .SelectMany(e => e.ConstructorArguments.Select(arg => arg.Value))
+            .OfType<string>()
+            .Where(e => !string.IsNullOrWhiteSpace(e))
+            .ToList();
+
+        //WriteMessage($"CustomAttribute: {isolatorCustomValues.Count} {string.Join(" ", isolatorCustomValues)}", MessageImportance.High);
+
+        //foreach (var attribute in isolatorCustomAttributes)
+        //{
+        //    WriteMessage($"CustomAttribute: {attribute} {string.Join(" ", attribute.ConstructorArguments.Select(e => e.Value))}", MessageImportance.High);
+        //}
+
+        return isolatorCustomValues;
+    }
+
+
+    private static void InjectMethodWithStringParameter(MethodDefinition method, MethodDefinition methodWithStringParameter, string value)
+    {
+        var il = method.Body.GetILProcessor();
+        var first = method.Body.Instructions.First();
+        InjectMethodWithStringParameter(il, first, methodWithStringParameter, value);
+    }
+    private static void InjectMethodWithStringParameter(ILProcessor il, Instruction first, MethodDefinition methodWithStringParameter, string value)
+    {
+        if (methodWithStringParameter is null)
+            return;
+
+        if (string.IsNullOrEmpty(value))
+        {
+            il.InsertBefore(first, il.Create(OpCodes.Ldnull));
+        }
+        else
+        {
+            il.InsertBefore(first, il.Create(OpCodes.Ldstr, value));
+        }
+
+        il.InsertBefore(first, il.Create(OpCodes.Call, methodWithStringParameter));
+    }
+
+}