+++ Isolator/IsolatorAttribute.cs
@@ -8,7 +8,13 @@ public class IsolatorAttribute : Attribute
     /// <summary>
     /// Initialize a new instance of <see cref="IsolatorAttribute"/>
     /// </summary>
-    public IsolatorAttribute()
+    public IsolatorAttribute(string contextName = null)
     {
+        ContextName = contextName;
     }
+
+    /// <summary>
+    /// Gets the name of the current context.
+    /// </summary>
+    public string ContextName { get; }
 }